+++ cypress/support/utils.js
@@ -1,0 +1,18 @@
+export const JOB_STATUSES = [
+    "WAITING",
+    "READY",
+    "PREPARED",
+    "SUBMITTED",
+    "HELD",
+    "QUEUING",
+    "RUNNING",
+    "COMPLETED",
+    "FAILED",
+    "SUSPENDED",
+    "SKIPPED",
+    "DELAYED"
+]
+
+export const DEFAULT_ITEMS_QUICK_VIEW = 100;
+
+export const DEFAULT_STATUS_QUICK_VIEW = "Any status";