+++ public/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## Changelog
 
-### Release v4.2.0 - Unreleased
+### Pre-release v4.2.0 - Unreleased
 * **Compatibility with Autosubmit API version >= [v4.2.0](https://pypi.org/project/autosubmit-api/v4.2.0/)** (unreleased)
+* Now the Quick View supports pagination.
 * Visual improvements (invert arrow direction in drop-down sections)
 
 ### Release v4.1.1 - Release date: 2026-09-21